some changes

+++ LandcapeGenVer1/HeightsMap/HeightsMap.h
@@ -21,7 +21,7 @@ public:
 
     void diamondSquare();
     void diamondSquare(float r);
-    void simpleGen(int r = 16, int n = 1);
+    void simpleGen(double r = 16, int n = 1);
 
     shared_ptr<HeightsMapPoints> createPoints(double kx, double ky, double kz);
     shared_ptr<HeightsMapPoints> createPoints();
@@ -32,7 +32,7 @@ private:
     void diamondSquare2(int x1, int y1, int x2, int y2, float range, unsigned level);
     double getRnd() const noexcept;
 
-    void randomizeHeightsMap(int x1, int y1, int x2, int y2, int n = 16) noexcept;
+    void randomizeHeightsMap(int x1, int y1, int x2, int y2, double n = 16) noexcept;
     void smoothHeightsMap(int x1, int y1, int x2, int y2, int n = 1) noexcept;
     double getValue(int i, int j);
 };
